Fix protocols in team export service [SCI-3037]
